Extract rules-list for future refactoring
