+++ src/docs.ts
@@ -1,7 +1,7 @@
 import {readFileSync, writeFileSync} from 'fs';
 import dedent from 'ts-dedent';
 import {DropdownOption} from './option';
-import {rules} from './rules';
+import {rules} from './rules-list';
 
 const autogen_warning = '<!--- This file was automatically generated. See docs.ts and *_template.md files for the source. -->\n';
